enterprise-rag-demo/requirements.txt
# Enterprise RAG Demo — Python 3.11+
openai>=1.12.0
faiss-cpu>=1.7.4
streamlit>=1.29.0
pypdf>=4.0.0
python-dotenv>=1.0.0
tiktoken>=0.5.0
